[FEAT] 로그인 유저 객체 가져오는 방식 변경 & jwt에 email 대신 userId 저장하도록 변경

+++ src/main/java/com/out4ider/selleing_backend/domain/novel/dto/NovelResponseDto.java
@@ -11,5 +11,5 @@ public class NovelResponseDto {
     private Long novelId;
     private String title;
     private String startSentence;
-    private int likeCount;
+    private long likeCount;
 }
